CmuxControlSocket: extract the listener transport + recovery policy from TerminalController
First stage of the CmuxControlSocket wave (modular refactor). New
Packages/CmuxControlSocket (depends on CmuxSocketControl) owns the
control-socket listener's path/bind/probe/lock machinery and its pure
recovery policy, lifted faithfully out of Sources/TerminalController.swift:

- SocketTransport (stateless Sendable struct): socket-path identity and
  liveness probing, advisory lock-file arbitration, listener binding,
  accepted-client configuration, writeAll, and the one-shot probeCommand
  client. Timeouts/backlog/lock vocabulary are stored config with the
  previous values as defaults.
- SocketListenerPolicy (stateless Sendable struct): accept-failure
  classification/backoff/rearm decisions, socket-path unlink rules, and
  the stable-default -> user-scoped bind-failure path fallback.
- Value types: SocketPathIdentity, SocketPathProbeResult,
  SocketBindAttemptResult, SocketStageFailure, SocketPathLockAcquisition,
  SocketAcceptErrorClassification (typed; raw values keep the existing
  breadcrumb strings), AcceptFailureRecoveryAction, SocketListenerHealth.

TerminalController keeps the listener loop/state and calls the package
through constructor-injected instances; tuple returns became value types
(SocketStageFailure, SocketPathLockAcquisition) with stage strings and
behavior unchanged. The accept-loop SocketControlServer actor is a later
stage. -782 lines from TerminalController (22,001 -> 21,219). Dead
socketProbePoll* constants deleted.

43 package tests (Swift Testing) migrated from cmuxTests or added:
accept-failure policy table, unlink rules, fallback table, path
identity/probe against real sockets, bind preparation, lock-file
arbitration incl. symlink rejection and reusable-marker round-trip,
writeAll, probeCommand timeout. cmuxTests keeps the start()/stop()
integration tests on the live listener. CI package-test list gains
CmuxControlSocket. TerminalControllerSocketWriteTests.swift renamed to
TerminalControllerTerminalTextTests.swift (its socket tests moved to the
package; only terminal-text tests remain).
