+++ .github/workflows/ci.yml
@@ -406,6 +406,7 @@ jobs:
           # app-target dependency). Add a package here once its `swift test`
           # is confirmed to resolve standalone.
           PACKAGES=(
+            CmuxControlSocket
             CmuxFileWatch
             CmuxFoundation
             CmuxGit
