+++ Sources/CmuxEventStream.swift
@@ -1,3 +1,4 @@
+import CmuxControlSocket
 import Darwin
 import Foundation
 
@@ -72,7 +73,7 @@ extension TerminalController {
 
     private nonisolated func writeEventsStreamLine(_ object: [String: Any], socket: Int32) -> Bool {
         guard let line = CmuxEventBus.encodeLine(object) else { return false }
-        return Self.writeAllToSocket(Data((line + "\n").utf8), to: socket)
+        return transport.writeAll(Data((line + "\n").utf8), to: socket)
     }
 
     private nonisolated static func stringSet(_ value: Any?) -> Set<String> {
