+++ cmuxTests/TerminalAndGhosttyTests.swift
@@ -1,5 +1,6 @@
 import XCTest
 import Testing
+import CmuxControlSocket
 import CmuxTerminalCopyMode
 import CmuxSocketControl
 import AppKit
@@ -5445,158 +5446,7 @@ final class GhosttyModifierFlagsChangedActionTests: XCTestCase {
 
 
 final class TerminalControllerSocketListenerHealthTests: XCTestCase {
-    func testStableSocketBindPermissionFailureFallsBackToUserScopedSocket() {
-        XCTAssertEqual(
-            TerminalController.fallbackSocketPathAfterBindFailure(
-                requestedPath: SocketControlSettings.stableDefaultSocketPath,
-                stage: "bind",
-                errnoCode: EACCES,
-                currentUserID: 501
-            ),
-            SocketControlSettings.userScopedStableSocketPath(currentUserID: 501)
-        )
-    }
-
-    func testNonStableSocketBindFailureDoesNotFallback() {
-        XCTAssertNil(
-            TerminalController.fallbackSocketPathAfterBindFailure(
-                requestedPath: "/tmp/cmux-debug.sock",
-                stage: "bind",
-                errnoCode: EACCES,
-                currentUserID: 501
-            )
-        )
-    }
-
-    func testStableSocketLockFailureFallsBackToUserScopedSocket() {
-        XCTAssertEqual(
-            TerminalController.fallbackSocketPathAfterBindFailure(
-                requestedPath: SocketControlSettings.stableDefaultSocketPath,
-                stage: "lock",
-                errnoCode: EWOULDBLOCK,
-                currentUserID: 501
-            ),
-            SocketControlSettings.userScopedStableSocketPath(currentUserID: 501)
-        )
-    }
-
-    func testStableSocketLockPreparationFailuresFallBackToUserScopedSocket() {
-        let failures: [(stage: String, errnoCode: Int32)] = [
-            ("create_lock_directory", EACCES),
-            ("open_lock", EACCES),
-            ("open_lock", ELOOP),
-            ("open_lock", EINVAL),
-            ("open_lock", EMLINK),
-        ]
-
-        for failure in failures {
-            XCTAssertEqual(
-                TerminalController.fallbackSocketPathAfterBindFailure(
-                    requestedPath: SocketControlSettings.stableDefaultSocketPath,
-                    stage: failure.stage,
-                    errnoCode: failure.errnoCode,
-                    currentUserID: 501
-                ),
-                SocketControlSettings.userScopedStableSocketPath(currentUserID: 501),
-                failure.stage
-            )
-        }
-    }
-
-    func testStableSocketExistingPathFailuresFallBackToUserScopedSocket() {
-        let failures: [(stage: String, errnoCode: Int32)] = [
-            ("existing_path", EEXIST),
-            ("stat_existing_path", EACCES),
-        ]
-
-        for failure in failures {
-            XCTAssertEqual(
-                TerminalController.fallbackSocketPathAfterBindFailure(
-                    requestedPath: SocketControlSettings.stableDefaultSocketPath,
-                    stage: failure.stage,
-                    errnoCode: failure.errnoCode,
-                    currentUserID: 501
-                ),
-                SocketControlSettings.userScopedStableSocketPath(currentUserID: 501),
-                failure.stage
-            )
-        }
-    }
-
-    func testSocketPathAcceptsConnectionsForLiveUnixSocket() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        defer {
-            Darwin.close(listenerFD)
-            unlink(path)
-        }
-
-        let handled = acceptSingleClient(on: listenerFD) { _ in }
-
-        XCTAssertTrue(TerminalController.socketPathAcceptsConnections(path))
-        wait(for: [handled], timeout: 1.0)
-    }
-
-    func testSocketPathAcceptsConnectionsRejectsStaleSocketFile() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        Darwin.close(listenerFD)
-        defer { unlink(path) }
-
-        XCTAssertFalse(TerminalController.socketPathAcceptsConnections(path))
-    }
-
-    func testPrepareSocketPathForBindRejectsRegularFileWithoutDeletingIt() throws {
-        let path = makeTempSocketPath()
-        try "not-a-socket".write(toFile: path, atomically: true, encoding: .utf8)
-        defer { unlink(path) }
-
-        let failure = try XCTUnwrap(TerminalController.prepareSocketPathForBind(path))
-
-        XCTAssertEqual(failure.stage, "existing_path")
-        XCTAssertEqual(failure.errnoCode, EEXIST)
-        XCTAssertTrue(FileManager.default.fileExists(atPath: path))
-    }
-
-    func testPrepareSocketPathForBindRejectsLiveSocketWithoutDeletingIt() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        defer {
-            Darwin.close(listenerFD)
-            unlink(path)
-        }
-
-        let handled = acceptSingleClient(on: listenerFD) { _ in }
-        let failure = try XCTUnwrap(TerminalController.prepareSocketPathForBind(path))
-
-        XCTAssertEqual(failure.stage, "bind")
-        XCTAssertEqual(failure.errnoCode, EADDRINUSE)
-        XCTAssertTrue(FileManager.default.fileExists(atPath: path))
-        wait(for: [handled], timeout: 1.0)
-    }
-
-    func testPrepareSocketPathForBindPreservesRefusedSocketFileWithoutPathLock() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        Darwin.close(listenerFD)
-        defer { unlink(path) }
-
-        let failure = try XCTUnwrap(TerminalController.prepareSocketPathForBind(path))
-
-        XCTAssertEqual(failure.stage, "bind")
-        XCTAssertEqual(failure.errnoCode, EADDRINUSE)
-        XCTAssertTrue(FileManager.default.fileExists(atPath: path))
-    }
-
-    func testPrepareSocketPathForBindRemovesRefusedSocketFileWithReusablePathLock() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        Darwin.close(listenerFD)
-        defer { unlink(path) }
-
-        XCTAssertNil(TerminalController.prepareSocketPathForBind(path, canReplaceRefusedSocket: true))
-        XCTAssertFalse(FileManager.default.fileExists(atPath: path))
-    }
+    private let transport = SocketTransport()
 
     @MainActor
     func testStartPreservesRefusedSocketFileWhenLockHasNoReusableMarker() throws {
@@ -5619,7 +5469,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
 
         XCTAssertTrue(FileManager.default.fileExists(atPath: path))
         XCTAssertTrue(FileManager.default.fileExists(atPath: path + ".lock"))
-        XCTAssertFalse(TerminalController.socketPathCanBeReclaimedForStartup(path))
+        XCTAssertFalse(transport.pathCanBeReclaimedForStartup(path))
         TerminalController.shared.start(
             tabManager: TabManager(),
             socketPath: path,
@@ -5627,7 +5477,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
         )
         XCTAssertTrue(FileManager.default.fileExists(atPath: path))
         XCTAssertTrue(FileManager.default.fileExists(atPath: path + ".lock"))
-        XCTAssertFalse(TerminalController.socketPathAcceptsConnections(path))
+        XCTAssertFalse(transport.pathAcceptsConnections(path))
     }
 
     @MainActor
@@ -5649,7 +5499,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
             accessMode: .allowAll
         )
 
-        XCTAssertTrue(TerminalController.socketPathAcceptsConnections(path))
+        XCTAssertTrue(transport.pathAcceptsConnections(path))
         XCTAssertTrue(FileManager.default.fileExists(atPath: path + ".lock"))
     }
 
@@ -5664,7 +5514,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
             socketPath: path,
             accessMode: .allowAll
         )
-        XCTAssertTrue(TerminalController.socketPathAcceptsConnections(path))
+        XCTAssertTrue(transport.pathAcceptsConnections(path))
 
         TerminalController.shared.stop()
         let listenerFD = try bindUnixSocket(at: path)
@@ -5673,7 +5523,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
             unlink(path)
             unlink(path + ".lock")
         }
-        XCTAssertTrue(TerminalController.socketPathCanBeReclaimedForStartup(path))
+        XCTAssertTrue(transport.pathCanBeReclaimedForStartup(path))
 
         TerminalController.shared.start(
             tabManager: TabManager(),
@@ -5681,7 +5531,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
             accessMode: .allowAll
         )
 
-        XCTAssertTrue(TerminalController.socketPathAcceptsConnections(path))
+        XCTAssertTrue(transport.pathAcceptsConnections(path))
     }
 
     @MainActor
@@ -5713,35 +5563,6 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
         XCTAssertFalse(FileManager.default.fileExists(atPath: path))
     }
 
-    func testStartupReclaimabilityAllowsMissingSocketWithoutLock() {
-        let path = makeTempSocketPath()
-        defer {
-            unlink(path)
-            unlink(path + ".lock")
-        }
-
-        XCTAssertTrue(TerminalController.socketPathCanBeReclaimedForStartup(path))
-    }
-
-    func testStartupReclaimabilityRejectsMissingSocketWithInvalidLock() throws {
-        let path = makeTempSocketPath()
-        let lockPath = path + ".lock"
-        let targetPath = path + ".target"
-        try "preserve me".write(toFile: targetPath, atomically: true, encoding: .utf8)
-        XCTAssertEqual(symlink(targetPath, lockPath), 0)
-        defer {
-            unlink(path)
-            unlink(lockPath)
-            unlink(targetPath)
-        }
-
-        XCTAssertFalse(TerminalController.socketPathCanBeReclaimedForStartup(path))
-        XCTAssertEqual(
-            try String(contentsOfFile: targetPath, encoding: .utf8),
-            "preserve me"
-        )
-    }
-
     @MainActor
     func testReservedStartupSocketPathFeedsActivePathBeforeListenerStarts() {
         TerminalController.shared.stop()
@@ -5766,7 +5587,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
             accessMode: .allowAll
         )
 
-        XCTAssertTrue(TerminalController.socketPathAcceptsConnections(reservedPath))
+        XCTAssertTrue(transport.pathAcceptsConnections(reservedPath))
     }
 
     @MainActor
@@ -5819,7 +5640,7 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
             socketPath: activePath,
             accessMode: .allowAll
         )
-        XCTAssertTrue(TerminalController.socketPathAcceptsConnections(activePath))
+        XCTAssertTrue(transport.pathAcceptsConnections(activePath))
 
         XCTAssertEqual(TerminalController.shared.reserveStartupSocketPath(reservedPath), reservedPath)
         XCTAssertFalse(FileManager.default.fileExists(atPath: reservedPath + ".lock"))
@@ -5882,32 +5703,6 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
         return fd
     }
 
-    private func acceptSingleClient(
-        on listenerFD: Int32,
-        handler: @escaping (_ clientFD: Int32) -> Void
-    ) -> XCTestExpectation {
-        let handled = expectation(description: "socket client handled")
-        DispatchQueue.global(qos: .userInitiated).async {
-            var clientAddr = sockaddr_un()
-            var clientAddrLen = socklen_t(MemoryLayout<sockaddr_un>.size)
-            let clientFD = withUnsafeMutablePointer(to: &clientAddr) { ptr in
-                ptr.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockaddrPtr in
-                    Darwin.accept(listenerFD, sockaddrPtr, &clientAddrLen)
-                }
-            }
-            guard clientFD >= 0 else {
-                handled.fulfill()
-                return
-            }
-            defer {
-                Darwin.close(clientFD)
-                handled.fulfill()
-            }
-            handler(clientFD)
-        }
-        return handled
-    }
-
     @MainActor
     func testSocketListenerHealthRecognizesSocketPath() throws {
         let path = makeTempSocketPath()
@@ -5934,91 +5729,4 @@ final class TerminalControllerSocketListenerHealthTests: XCTestCase {
         XCTAssertFalse(health.isHealthy)
     }
 
-    func testProbeSocketCommandReturnsFirstLineResponse() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        defer {
-            Darwin.close(listenerFD)
-            unlink(path)
-        }
-
-        let handled = acceptSingleClient(on: listenerFD) { clientFD in
-            var buffer = [UInt8](repeating: 0, count: 256)
-            _ = read(clientFD, &buffer, buffer.count)
-            let response = "PONG\nextra\n"
-            _ = response.withCString { ptr in
-                write(clientFD, ptr, strlen(ptr))
-            }
-        }
-
-        let response = TerminalController.probeSocketCommand("ping", at: path, timeout: 0.5)
-
-        XCTAssertEqual(response, "PONG")
-        wait(for: [handled], timeout: 1.0)
-    }
-
-    func testProbeSocketCommandTimesOutWithoutPollingUntilServerResponds() throws {
-        let path = makeTempSocketPath()
-        let listenerFD = try bindUnixSocket(at: path)
-        defer {
-            Darwin.close(listenerFD)
-            unlink(path)
-        }
-
-        let releaseServer = DispatchSemaphore(value: 0)
-        let handled = acceptSingleClient(on: listenerFD) { clientFD in
-            var buffer = [UInt8](repeating: 0, count: 256)
-            _ = read(clientFD, &buffer, buffer.count)
-            _ = releaseServer.wait(timeout: .now() + 1.0)
-        }
-
-        let startedAt = Date()
-        let response = TerminalController.probeSocketCommand("ping", at: path, timeout: 0.2)
-        let elapsed = Date().timeIntervalSince(startedAt)
-        releaseServer.signal()
-
-        XCTAssertNil(response)
-        XCTAssertGreaterThanOrEqual(elapsed, 0.18)
-        XCTAssertLessThan(elapsed, 0.8)
-        wait(for: [handled], timeout: 1.0)
-    }
-
-    func testSocketListenerHealthFailureSignalsAreEmptyWhenHealthy() {
-        let health = TerminalController.SocketListenerHealth(
-            isRunning: true,
-            acceptLoopAlive: true,
-            socketPathMatches: true,
-            socketPathExists: true,
-            socketPathOwnedByListener: true
-        )
-        XCTAssertTrue(health.isHealthy)
-        XCTAssertTrue(health.failureSignals.isEmpty)
-    }
-
-    func testSocketListenerHealthFailureSignalsIncludeAllDetectedProblems() {
-        let health = TerminalController.SocketListenerHealth(
-            isRunning: false,
-            acceptLoopAlive: false,
-            socketPathMatches: false,
-            socketPathExists: false,
-            socketPathOwnedByListener: false
-        )
-        XCTAssertFalse(health.isHealthy)
-        XCTAssertEqual(
-            health.failureSignals,
-            ["not_running", "accept_loop_dead", "socket_path_mismatch", "socket_missing"]
-        )
-    }
-
-    func testSocketListenerHealthReportsIdentityMismatchSeparately() {
-        let health = TerminalController.SocketListenerHealth(
-            isRunning: true,
-            acceptLoopAlive: true,
-            socketPathMatches: true,
-            socketPathExists: true,
-            socketPathOwnedByListener: false
-        )
-        XCTAssertFalse(health.isHealthy)
-        XCTAssertEqual(health.failureSignals, ["socket_identity_mismatch"])
-    }
 }
